% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_xdg.R
\name{path_xdg_cache_home}
\alias{path_xdg_cache_home}
\alias{path_xdg_config_home}
\alias{path_xdg_data_home}
\alias{path_xdg_state_home}
\alias{path_xdg_runtime_dir}
\title{XDG Base Directory Paths}
\usage{
path_xdg_cache_home(...)

path_xdg_config_home(...)

path_xdg_data_home(...)

path_xdg_state_home(...)

path_xdg_runtime_dir(...)
}
\arguments{
\item{...}{Additional paths appended to the directory by \link[fs:path]{fs::path}}
}
\value{
A string containing a path
}
\description{
These functions allow you to construct paths that adhere to the
\href{https://specifications.freedesktop.org/basedir-spec/basedir-spec-latest.html}{XDG Base Directory Specification}.?

\code{path_xdg_cache_home()} constructs a path beginning at the
directory in which user-specific non-essential data files should be stored.

\code{path_xdg_config_home()} constructs a path beginning at the
directory in which user-specific configuration files should be stored.

\code{path_xdg_data_home()} constructs a path beginning at the
directory in which user-specific data files should be stored.

\code{path_xdg_state_home()} constructs a path beginning at the
directory in which user-specific state files should be stored.

\code{path_xdg_runtime_dir()} constructs a path beginning at the
directory in which user-specific runtime files and other file objects should
be placed. If \verb{$XDG_RUNTIME_DIR} is not set, \code{NA} is returned.
}
\examples{
path_xdg_cache_home()
path_xdg_config_home()
path_xdg_data_home()
path_xdg_state_home()
path_xdg_state_home()
}
