% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xdg_is_set.R
\name{xdg_cache_home_is_set}
\alias{xdg_cache_home_is_set}
\alias{xdg_config_home_is_set}
\alias{xdg_data_home_is_set}
\alias{xdg_state_home_is_set}
\title{Check whether an XDG Base Directory environment variable is set}
\usage{
xdg_cache_home_is_set()

xdg_config_home_is_set()

xdg_data_home_is_set()

xdg_state_home_is_set()
}
\value{
A logical value
}
\description{
Check whether an XDG Base Directory environment variable is set
}
